api/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.4
django-hashid-field==3.3.4
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gunicorn==20.1.0
hashids==1.3.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.2.0
protobuf==4.21.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rest-framework-generic-relations==2.1.0
rsa==4.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
